requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
numpy
scikit-learn
seaborn
matplotlib
xgboost
db-dtypes
google-cloud-bigquery-storage